Patricio

+++ src/controller/courseController.js
@@ -26,7 +26,7 @@ exports.create = (req, res) => {
   const data = req.body;
   Restaurant.findById(data.Restaurant)
   .then(restaurant => {
-    if(!data.name) return res.status(500).json({Message: "Missing name of Course",});
+    if(!data.name) return Promise.reject('Missing name of Course');
     const newCourse = new Course({
       name: data.name, 
       Restaurant: restaurant._id,
@@ -34,7 +34,7 @@ exports.create = (req, res) => {
     newCourse.save();
   })
   .then((newCourse)=> { res.status(201).json({Message: "Your new course was created Succesfully", newCourse})})
-  .catch(error => { res.status(500).json(error)});
+  .catch(error => { res.status(500).json({message: error})});
 
 }
 
@@ -50,22 +50,30 @@ exports.deleteCourse = (req,res) => {
 }
 exports.deleteCoursesAndDishes = (req,res) => {
     const id = req.params.id;
-    Course.findById(id).then((course)=>{
-      Dish.deleteMany({ Course: course._id },(err,dishes)=>{
-        if(err) {
-          return res.status(500).json({message: err.getMessage()})
-        }
-        console.log(`The Dishes from the Course ${course.name} have been removed`)
-      });
-    });
     Course.findByIdAndDelete(id)
+    .then((course)=>{
+            if(course === null ) return Promise.reject('Course not Found');
+            Dish.deleteMany({ Course: course._id },(err,dishes)=>{
+                if(err) return console.log("Problem trying to delete the dishes");
+            })
+        })
     .then(() => {
-      res.status(200).json({message: "Your Course has been deleted Succesfully"})
+      res.status(200).json({
+        confirmation: "Success",
+        message: `The dishes from the Course ${id} have been deleted` ,
+        })
     })
     .catch(error=>{
-      res.status(500).json(error);
+        if(error === 'Course not Found') {
+            return res.status(404).json({message: error});
+        } else {
+            return res.status(500).json({message: error});
+        }
+      
     })
-  }
+    }
+
+  
 
 exports.update = (req,res) => {
   const id = req.params.id;
@@ -91,4 +99,19 @@ exports.search = (req, res) => {
   .catch(error => {
     res.status(500).json(error);
   });
+}
+
+exports.coursesWithDishes = (req, res) => {
+    const id = req.params.id;
+
+    Restaurant.findById(id)
+    .then(rest => {
+        if (!rest) return Promise.reject('Problem finding restaurant');
+        Course.find({Restaurant: rest._id}).populate('dishList').exec((error, dishes) => {
+            if (error) return Promise.reject('Problem populating courses');;
+            res.status(200).json(dishes);
+          })
+    })
+    
+    .catch(error => { res.status(500).json({message: error})});
 }
